+++ src/helpers/util.ts
@@ -6,6 +6,10 @@ const util = {
             return (number / 1000).toFixed(1) + 'K';
         }
         return '' + number;
+    },
+
+    currentUtc() {
+        return new Date().getTime();
     }
 }
